Take apart long backend spec file
